# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BASE",
        "description": "Servicio API REST para la gestión y administración del sistema Automatizado de riego por goteo",
        "contact": {
            "name": "SETI",
            "email": "cirojas@cenicana.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/v1/drip-irrigation-modules": {
            "get": {
                "tags": [
                    "DripIrrigationModules"
                ],
                "summary": "Get list of drip irrigation modules",
                "description": "Returns list of drip irrigation modules ",
                "operationId": "getDripIrrigationModuleList",
                "parameters": [
                    {
                        "name": "estate_irrigation_system_id",
                        "in": "query",
                        "description": "The ID of the estate irrigation system ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DripIrrigationModules"
                ],
                "summary": "Store new drip irrigation module",
                "description": "Returns drip irrigation module data",
                "operationId": "storeDripIrrigationModule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DripIrrigationModule"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DripIrrigationModule"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/drip-irrigation-modules/{id}": {
            "get": {
                "tags": [
                    "DripIrrigationModules"
                ],
                "summary": "Get drip irrigation module information",
                "description": "Returns drip irrigation module data",
                "operationId": "getDripIrrigationModuleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Drip irrigation module id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/dropper-types": {
            "get": {
                "tags": [
                    "DropperTypes"
                ],
                "summary": "Get list of dropper types",
                "description": "Returns list of dropper types",
                "operationId": "getDropperTypesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/surco-separations": {
            "get": {
                "tags": [
                    "SurcoSeparations"
                ],
                "summary": "Get list of surco separations ",
                "description": "Returns list of surco separations (mts)",
                "operationId": "getSurcoSeparationsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "DripIrrigationModule": {
                "title": "Drip Irrigation  model",
                "description": "DripIrrigationModule model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the irrigation system (min: 3 characters)",
                        "type": "string",
                        "example": "Sistema A"
                    },
                    "area": {
                        "title": "Area",
                        "description": "Area of the module ",
                        "type": "number",
                        "format": "float",
                        "example": "60"
                    },
                    "estate_irrigation_system_id": {
                        "title": "EstateIrrigationSystemId",
                        "description": "The ID of the irrigation system (sistema de riego de una hacienda)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "dropper_type_id": {
                        "title": "DropperTypeId",
                        "description": "The ID of the dropper type (el tipo de un gotero)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "dropper_flow": {
                        "title": "Dropper flow",
                        "description": "The flow of the dropper (flujo del gotero)",
                        "type": "number",
                        "format": "float",
                        "example": "2.0"
                    },
                    "surco_separation_id": {
                        "title": "SurcoSeparationId",
                        "description": "The ID of the surco separation (separación entre surcos)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "dropper_separation": {
                        "title": "Dropper separation",
                        "description": "The separation between droppers (separación entre goteros mts)",
                        "type": "number",
                        "format": "float",
                        "example": "0.5"
                    },
                    "irrigation_strategy_id": {
                        "title": "IrrigationStrategyId",
                        "description": "The ID of the irrigation strategy (estrategia de riego)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "irrigation_amount_minutes": {
                        "title": "IrrigationMinutes",
                        "description": "Minutes for the irrigation (minutos de riego)",
                        "type": "integer",
                        "format": "int64",
                        "example": "80"
                    },
                    "irrigation_frequency_days": {
                        "title": "IrrigationDays",
                        "description": "The amount of irrigation days per week (# días de riego a la semana)",
                        "type": "integer",
                        "format": "int64",
                        "example": "5"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DripIrrigationModule"
                }
            },
...
            "EntityJsonResource": {
                "title": "EntityJsonResource",
                "description": "A Json wrapper resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EntityJsonResource"
                }
            }
        },
        "securitySchemes": {
            "jwt_key": {
                "type": "apiKey",
                "description": "Debe insertar el token generado en el proceso de autenticación",
                "name": "Authorization",
                "in": "header",
                "template": "Bearer {}"
            }
        }
    },
...
}